Wiat for DOM load and check for > IE 9
``#appendData`` gives an error ``undefined is not a function`` and doesn't load the script. 

This new approach creates the IE conditional comment in memory and only builds / embeds the shim if the browser lets us select nodes inside the conditional. 

The ``Meteor.isClient`` scopes this to only run in the browser.

The ``Meteor.startup`` ensures the document is ready before we inject the js.